+++ app/js/controllers/admin_controller.js
@@ -3,56 +3,37 @@ angular.module('Dashboard.Admin', [])
 .controller('AddCtrl', ['$scope', '$http', 'Admin', 'Sources',
     function($scope, $http, Admin, Sources) {
 
-        // List of possible templates
-        // >>>>>>>>>>>>>>>>>>>>>>>>>>
-        $scope.dbWidgets = {
-            "linechart": {
-                "points": 0
-            },
-            "delta": {},
-            "gauge": {
-                "min": 0,
-                "max": 0,
-            },
-            "sum": {
-                "append": "",
-                "prepend": "",
-                "subtitle": ""
-            },
-            "list": {
-                "limit": 0
-            },
-            "picture": {},
-            "gauge": {},
-            "countdown": {
-                "startdate": 0,
-                "enddate": 0
-            },
-            "announcement": {
-                "announcement": "",
-                "subtitle": ""
-            }
-        }
-
         // Get list of useable datasources
         // >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
-        $scope.sources = Sources.getSources();
+        Sources.getSources().then(function(data) {
+            $scope.sources = data;
+        });
+
+        // Get templates in /js/config/config.json
+        // >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+        Sources.getTemplates().then(function(data) {
+            $scope.dbWidgets = data[0];
+        });
 
         // Template Options
         // >>>>>>>>>>>>>>>>
         $scope.template1 = [];
         $scope.template2 = [];
 
-        $scope.template1Option = function(template1) {
+        $scope.template1Add = function(template1) {
             $scope.template1.push({"key": template1.key, "value": template1.value});
-            console.log($scope.template1);
             template1.key = "", template1.value = "";
         }
+        $scope.template1Remove = function(index) {
+            $scope.template1.splice(index, 1);
+        }
 
-        $scope.template2Option = function(template2) {
+        $scope.template2Add = function(template2) {
             $scope.template2.push({"key": template2.key, "value": template2.value});
-            console.log($scope.template2)
             template2.key = "", template2.value = "";
+        }
+        $scope.template2Remove = function(index) {
+            $scope.template2.splice(index, 1);
         }
 
         // Add widget function
@@ -68,7 +49,6 @@ angular.module('Dashboard.Admin', [])
             // Construct Widget
             // >>>>>>>>>>>>>>>>
             $scope.realWidget = {
-                "_id" : widget._id,
                 "config": {
                     "refresh": widget.config.refresh,
                     "source": widget.config.source,
@@ -85,6 +65,12 @@ angular.module('Dashboard.Admin', [])
                 }
             }
 
+            // Add ID if not random
+            // >>>>>>>>>>>>>>>>>>>>
+            if ($scope.randomID === false) {
+                $scope.realWidget._id = widget._id;
+            }
+
             // Check Source Value
             // >>>>>>>>>>>>>>>>>>
             if(widget.config.source == undefined) {
@@ -96,20 +82,21 @@ angular.module('Dashboard.Admin', [])
             for(i = 0; i < $scope.template1.length; i++) {
                 dir1[$scope.template1[i].key] = $scope.template1[i].value;
             };
-
-            for(i = 0; i < $scope.template2.length; i++) {
-                dir2[$scope.template2[i].key] = $scope.template2[i].value;
-            };
-
             $scope.realWidget.content.templates[temp1] = dir1;
-            $scope.realWidget.content.templates[temp2] = dir2;
+            
+            // If only 1 template, don't add 2nd as undefined
+            if (temp2 != undefined || temp2 != null ) {
+                for(i = 0; i < $scope.template2.length; i++) {
+                    dir2[$scope.template2[i].key] = $scope.template2[i].value;
+                };
+                $scope.realWidget.content.templates[temp2] = dir2;
+            }
 
             console.log($scope.realWidget);
 
             // POST to Database
             // >>>>>>>>>>>>>>>>
             Admin.addWidget($scope.realWidget);
-
         };
 
     }
@@ -117,69 +104,55 @@ angular.module('Dashboard.Admin', [])
 
 .controller('UpdateCtrl', ['$scope', '$http', '$location', 'Admin', 'Sources', 'Widgets',
     function($scope, $http, $location, Admin, Sources, Widgets) {
-
-        // List of possible templates
-        // >>>>>>>>>>>>>>>>>>>>>>>>>>
-        $scope.dbWidgets = {
-            "linechart": {
-                "points": 0
-            },
-            "delta": {},
-            "gauge": {
-                "min": 0,
-                "max": 0,
-            },
-            "sum": {
-                "append": "",
-                "prepend": "",
-                "subtitle": ""
-            },
-            "list": {
-                "limit": 0
-            },
-            "picture": {},
-            "gauge": {},
-            "countdown": {
-                "startdate": 0,
-                "enddate": 0
-            },
-            "announcement": {
-                "announcement": "",
-                "subtitle": ""
-            }
-        }
-
-        // Get list of useable datasources
-        // >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
-        $scope.sources = Sources.getSources();
+        $scope.updateSource = false;
+        $scope.updateTemplates = false;
 
         // Get widget ID
         // >>>>>>>>>>>>>
         var id = $location.hash();
-        $scope.thisId = id;
+        var thisId = id;
 
         // Get the widget to edit
         // >>>>>>>>>>>>>>>>>>>>>>
-        Widgets.getWidgetById($scope.thisId).then(function(data) {
+        Widgets.getWidgetById(thisId).then(function(data) {
             $scope.widget = data;
+            // $scope.widget.config.source = data.config.source;
+            console.log(data.config.source);
+            $scope.thisTitle = data.content.title;
         });
 
+        // Get list of useable datasources
+        // >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+        Sources.getSources().then(function(data) {
+            $scope.sources = data;
+        });
+
+        // Get templates in /js/config/config.json
+        // >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+        Sources.getTemplates().then(function(data) {
+            $scope.dbWidgets = data[0];
+        });
 
         // Template Options
         // >>>>>>>>>>>>>>>>
         $scope.template1 = [];
         $scope.template2 = [];
 
-        $scope.template1Option = function(template1) {
+        $scope.template1Add = function(template1) {
             $scope.template1.push({"key": template1.key, "value": template1.value});
             template1.key = "", template1.value = "";
         }
+        $scope.template1Remove = function(index) {
+            $scope.template1.splice(index, 1);
+        }
 
-        $scope.template2Option = function(template2) {
+        $scope.template2Add = function(template2) {
             $scope.template2.push({"key": template2.key, "value": template2.value});
             template2.key = "", template2.value = "";
         }
-
+        $scope.template2Remove = function(index) {
+            $scope.template2.splice(index, 1);
+        }
 
         // Update Function
         // >>>>>>>>>>>>>>>