+++ app/js/directives/charts_directives.js
@@ -11,43 +11,64 @@ angular.module('Dashboard.Charts', [])
         },
         controller: function($scope, $element, $timeout) {
             $timeout(function () {
-                var axisData = JSON.parse($scope.data),
-                    width = $element.width(),
-                    height = $element.height(),
-                    parseDate = d3.time.format("%d-%b-%y").parse, // date format like this '28-Mar-12'
-                    x = d3.time.scale().range([0, width]),
-                    y = d3.scale.linear().range([height, 0]),
-                    xAxis = d3.svg.axis().scale(x).orient("bottom"),
-                    yAxis = d3.svg.axis().scale(y).orient("left"),
-                    line = d3.svg.line()
-                        .x(function(d) { 
-                            return x(d.value.date); 
-                        })
-                        .y(function(d) { 
-                            return y(d.value.amount); 
-                        }),
-                    svg = d3.select($element[0]).append("svg").attr("width", width).attr("height", height).append("g");
+                var linechart = function() {
+                    // Remove old graph, replace with new graph
+                    angular.element($element[0]).children().remove();
 
-                angular.forEach(axisData, function(data, index, source){
-                    var target = data.value;
+                    var axisData = JSON.parse($scope.data),
+                        width = $element.width(),
+                        height = $element.height(),
+                        parseDate = d3.time.format("%d-%b-%y").parse, // date format like this '28-Mar-12'
+                        x = d3.time.scale().range([0, width]),
+                        y = d3.scale.linear().range([height, 0]),
+                        xAxis = d3.svg.axis().scale(x).orient("bottom"),
+                        yAxis = d3.svg.axis().scale(y).orient("left"),
+                        line = d3.svg.line()
+                            .x(function(d) { 
+                                return x(d.value.date); 
+                            })
+                            .y(function(d) { 
+                                return y(d.value.amount);
+                            }),
+                        area = d3.svg.area()
+                            .x(function(d) {
+                                return x(d.value.date);
+                            })
+                            .y0(height)
+                            .y1(function(d) {
+                                return y(d.value.amount);
+                            }),
+                        svg = d3.select($element[0]).append("svg").attr("width", width).attr("height", height).append("g");
 
-                    target.date = parseDate(target.date);
-                    target.amount = +target.amount.toFixed();
-                });
+                    
 
-                // sort by date, its important, linechart need
-                axisData.sort(function (x, y) { return x['value']['date'] - y['value']['date'] })
+                    angular.forEach(axisData, function(data, index, source){
+                        var target = data.value;
 
-                x.domain(d3.extent(axisData, function(d) { return d.value.date; }));
-                y.domain(d3.extent(axisData, function(d) { return d.value.amount; }));
+                        target.date = parseDate(target.time);
+                        target.amount = +target.data.value.toFixed(); //target.amount
+                    });
 
-                // line
-                svg.append("path").datum(axisData).attr("class", "line").attr("d", line);
+                    // sort by date, its important, linechart need
+                    axisData.sort(function (x, y) { return x['value']['date'] - y['value']['date'] })
+
+                    x.domain(d3.extent(axisData, function(d) { return d.value.date; }));
+                    y.domain(d3.extent(axisData, function(d) { return d.value.amount; }));
+
+                    // line
+                    svg.append("path").datum(axisData).attr("class", "area").attr("d", area);
+                    svg.append("path").datum(axisData).attr("class", "line").attr("d", line);
+                }
+
+                $scope.$watch('data', function (aft, bef) {
+                    linechart();
+                }, true);
 
             }, 0)
         }
     };
 })
+
 
 // Countdown Directive
 .directive('countdown', function() {
@@ -127,42 +148,49 @@ angular.module('Dashboard.Charts', [])
         },
         templateUrl: 'templates/delta.html',
         controller: function($scope, $element) {
-            var sData = JSON.parse($scope.data),
-                parseDate = d3.time.format("%d-%b-%y").parse; // Correct time formatting
-            
-            angular.forEach(sData, function(data, index, source) {
-                var target = data.value;
-                target.date = parseDate(target.date);
-            });
-            
-            // Order Information by date
-            sData.sort(function(a, b) { return a['value']['date'] - b['value']['date'] });
+            var delta = function() {
+                var sData = JSON.parse($scope.data),
+                    parseDate = d3.time.format("%d-%b-%y").parse; // Correct time formatting
+                
+                angular.forEach(sData, function(data, index, source) {
+                    var target = data.value;
+                    target.date = parseDate(target.time);
+                });
+                
+                // Order Information by date
+                sData.sort(function(a, b) { return a['value']['date'] - b['value']['date'] });
 
-            var base = sData[0].value.amount,
-                length = sData.length - 1,
-                current = sData[length].value.amount,
-                diff = ((current / base) * 100);
+                var base = sData[0].value.data.value,
+                    length = sData.length - 1,
+                    current = sData[length].value.data.value,
+                    diff = ((current / base) * 100);
 
-            var status;
-            if (diff > 100) {
-                dir = "⬆";
-                status = 'up';
-            } else if (diff === 100) {
-                dir = "="
-                status = 'equal'
-            } else {
-                dir = "⬇";
-                status = 'down'
+                var status;
+                if (diff > 100) {
+                    dir = "⬆";
+                    status = 'up';
+                } else if (diff === 100) {
+                    dir = "="
+                    status = 'equal'
+                } else {
+                    dir = "⬇";
+                    status = 'down'
+                }
+
+                diff = diff - 100;
+                diff = diff.toFixed();
+
+                $scope.delta = {
+                    'difference' : diff + "%",
+                    'direction' : dir,
+                    'status' : status
+                }                
             }
 
-            diff = diff - 100;
-            diff = diff.toFixed();
+            $scope.$watch('data', function (aft, bef) {
+                delta();
+            }, true); 
 
-            $scope.delta = {
-                'difference' : diff + "%",
-                'direction' : dir,
-                'status' : status
-            }
         }
     };
 })
@@ -178,20 +206,27 @@ angular.module('Dashboard.Charts', [])
         },
         templateUrl: 'templates/sum.html',
         controller: function($scope, $element, $timeout, Number) {
-            var sData = JSON.parse($scope.data),
-                tmpls = JSON.parse($scope.templates),
-                append = tmpls.sum.append,
-                prepend = tmpls.sum.prepend,
-                sub = tmpls.sum.subtitle,
-                total = 0;
+            var sum = function() {
+                var sData = JSON.parse($scope.data),
+                    tmpls = JSON.parse($scope.templates),
+                    append = tmpls.sum.append,
+                    prepend = tmpls.sum.prepend,
+                    sub = tmpls.sum.subtitle,
+                    total = 0;
 
-            angular.forEach(sData, function(data, index, source){
-                total += data.value.amount;
-            });
+                angular.forEach(sData, function(data, index, source){
+                    total += data.value.data.value;
+                });
 
-            total = Number(total, prepend, append);
-            $scope.sum = total;
-            $scope.sub = sub;
+                total = Number(total, prepend, append);
+                $scope.sum = total;
+                $scope.sub = sub;
+            }
+
+            $scope.$watch('data', function (aft, bef) {
+                sum();
+            }, true);
+
         }
     };
 })
@@ -207,30 +242,37 @@ angular.module('Dashboard.Charts', [])
         },
         templateUrl: 'templates/list.html',
         controller: function($scope, $element, $timeout) {
-            var sData = JSON.parse($scope.data),
+            var list = function() {
+                var sData = JSON.parse($scope.data),
                 tmpls = JSON.parse($scope.templates).list,
                 nLimit = parseInt(tmpls.limit);
 
-            var array = [];
-            angular.forEach(sData, function(data, index) {
-                if (data.value.time !== undefined) {
-                    data.value.time = moment(data.value.time).fromNow();
-                };
-                array.push(data.value);
-            });
+                var array = [];
+                angular.forEach(sData, function(data, index) {
+                    if (data.value.time !== undefined) {
+                        data.value.time = moment(data.value.time).fromNow();
+                    };
+                    array.push(data.value);
+                });
 
-            // Show status icons if builds, img if not
-            $scope.builds = false;
-            if (sData[0].value.status !== undefined) {
-                $scope.builds = true;
+                // Show status icons if builds, img if not
+                $scope.builds = false;
+                if (sData[0].value.data.status !== undefined) {
+                    $scope.builds = true;
+                };
+
+                // Data into scope
+                if (nLimit) {   // if has limit attr
+                    $scope.list = array.splice(0, nLimit);
+                } else {
+                    $scope.list = array;
+                }
             };
 
-            // Data into scope
-            if (nLimit) {   // if has limit attr
-                $scope.list = array.splice(0, nLimit);
-            } else {
-                $scope.list = array;
-            }
+            $scope.$watch('data', function() {
+                list();
+            }, true);
+
         }
     }
 })
@@ -264,75 +306,313 @@ angular.module('Dashboard.Charts', [])
         },
         templateUrl: 'templates/picture.html',
         controller: function($scope, $element) {
-            var sData = JSON.parse($scope.data),
-                image = sData[0].value.image;
+            var picture = function() {
+                var sData = JSON.parse($scope.data),
+                image = sData[0].value.data.image;
 
-            $scope.imageUrl = image;
+                $scope.imageUrl = image;
+            }
+
+            $scope.$watch('data', function() {
+                picture();
+            }, true);
 
         }
     }
 })
 
-// Gauge
+// Clock directive
+.directive('clock', function() {
+    return {
+        restrict: 'E',
+        replace: true,
+        scope: {
+        },
+        controller: function($scope, $element, $timeout) {
+            $timeout(function () {
+                var fields, width, height, offSetX, offSetY, pi, scaleSecs, scaleMins, scaleHours, vis, clockGroup, render;
+
+                fields = function() {
+                    var currentTime, hour, minute, second;
+                    currentTime = new Date();
+                    second = currentTime.getSeconds();
+                    minute = currentTime.getMinutes() + second / 60;
+                    hour = currentTime.getHours() + minute / 60;
+                    return data = [
+                        {
+                            "unit": "seconds",
+                            "numeric": second
+                        },
+                        {
+                            "unit": "minutes",
+                            "numeric": minute
+                        },
+                        {
+                            "unit": "hours",
+                            "numeric": hour
+                        }
+                    ];
+                };
+
+                // config
+                width = $element.width();
+                height = $element.height();
+                offSetX = 125;
+                offSetY = 100;
+
+                pi = Math.PI;
+                scaleSecs = d3.scale.linear().domain([0, 59 + 999/1000]).range([0, 2 * pi]);
+                scaleMins = d3.scale.linear().domain([0, 59 + 59/60]).range([0, 2 * pi]);
+                scaleHours = d3.scale.linear().domain([0, 11 + 59/60]).range([0, 2 * pi]);
+
+                // Append clockGroup
+                vis = d3.select($element[0])
+                    .append("svg:svg")
+                    .attr("width", width)
+                    .attr("height", height);
+
+                clockGroup = vis.append("svg:g")
+                    .attr("transform", "translate(" + offSetX + "," + offSetY + ")");
+
+                clockGroup.append("svg:circle")
+                    .attr("r", 80).attr("fill", "none")
+                    .attr("class", "clock outercircle")
+                    .attr("stroke", "rgba(184, 179, 255, 0.5)")
+                    .attr("stroke-width", 2);
+
+                clockGroup.append("svg:circle")
+                    .attr("r", 4)
+                    .attr("fill", "#b8b3ff")
+                    .attr("class", "clock innercircle");
+
+                // render clockHands
+                render = function(data) {
+                    var hourArc, minuteArc, secondArc;
+
+                    clockGroup.selectAll(".clockhand").remove();
+
+                    secondArc = d3.svg.arc().innerRadius(0).outerRadius(70)
+                        .startAngle(function(d) { return scaleSecs(d.numeric); })
+                        .endAngle(function(d) { return scaleSecs(d.numeric); });
+
+                    minuteArc = d3.svg.arc().innerRadius(0).outerRadius(70)
+                        .startAngle(function(d) { return scaleMins(d.numeric); })
+                        .endAngle(function(d) { return scaleMins(d.numeric); });
+
+                    hourArc = d3.svg.arc().innerRadius(0).outerRadius(50)
+                        .startAngle(function(d) { return scaleHours(d.numeric % 12); })
+                        .endAngle(function(d) { return scaleHours(d.numeric % 12); });
+
+                    clockGroup.selectAll(".clockhand")
+                        .data(data)
+                        .enter()
+                        .append("svg:path")
+                        .attr("d", function(d) {
+                            if (d.unit === "seconds") {
+                                return secondArc(d);
+                            } else if (d.unit === "minutes") {
+                                return minuteArc(d);
+                            } else if (d.unit === "hours") {
+                                return hourArc(d);
+                            }
+                        })
+                        .attr("class", "clockhand")
+                        .attr("stroke", "#b8b3ff")
+                        .attr("stroke-width", function(d) {
+                            if (d.unit === "seconds") {
+                                return 2;
+                            } else if (d.unit === "minutes") {
+                                return 3;
+                            } else if (d.unit === "hours") {
+                                return 3;
+                            }
+                        })
+                        .attr("fill", "none");
+                };
+
+                // render the clock
+                render(fields());
+                // and keep it ticking
+                setInterval(function() {
+                    render(fields());
+                }, 1000);
+
+            }, 0) //$timeout
+        }
+    };
+})
+
+//Gauge directive
 .directive('gauge', function() {
     return {
         restrict: 'E',
         replace: true,
         scope: {
-            templates: '@',
-            data: '@'
+            data: '@',
+            templates: '@'
         },
-        templateUrl: "templates/gauge.html",
-        controller: function($scope, $element) {
-            var sData = JSON.parse($scope.data),
-                tmpls = JSON.parse($scope.templates);
+        controller: function($scope, $element, $timeout) {
+            $timeout(function () {
+                var powerGauge,
+                    sData = JSON.parse($scope.data),
+                    tmpls = JSON.parse($scope.templates),
+                    gauge = function(configuration) {
+                        var that = {},
+                            config = {
+                                size                     : ($element.width() - 20),
+                                clipWidth                : $element.width(),
+                                clipHeight               : $element.height()/1.5,
+                                ringInset                : 20,
+                                ringWidth                : 40,
+                                pointerWidth             : 10,
+                                pointerTailLength        : 5,
+                                pointerHeadLengthPercent : 0.9,
+                                
+                                minValue                 : tmpls.gauge.min || 0,
+                                maxValue                 : tmpls.gauge.max || 1000,
 
-            var svg = d3.select($element[0]).append("svg");
-            var pi = Math.PI;
+                                minAngle                 : -90,
+                                maxAngle                 : 90, 
 
-            var min = parseFloat(tmpls.gauge.min);
-            var max = parseFloat(tmpls.gauge.max);
-            var current = 589;
+                                transitionMs             : 3500,
+                                majorTicks               : 5,
+                                labelFormat              : d3.format(',g'),
+                                labelInset               : 10,
+                                arcColorFn               : d3.interpolateHsl(d3.rgb('#AEAEFF'), d3.rgb('#383872'))
+                            },
+                            range, r, pointerHeadLength, svg, arc, scale, ticks, tickData, pointer,
+                            value = 0,
+                            donut = d3.layout.pie(),
+                            deg2rad = function(deg) {
+                                return deg * Math.PI / 180;
+                            },
+                            newAngle = function(d) {
+                                var newAngle = config.minAngle + (scale(d) * range);
+                                return newAngle;
+                            },
+                            configure = function(configuration) {
+                                var prop = undefined;
 
-            // No higher/lower than max/min
-            if (current > max) {
-                current = max;
-            }
-            if (current < min) {
-                current = min;
-            }
+                                for ( prop in configuration ) {
+                                    config[prop] = configuration[prop];
+                                }
 
-            // Color Scheme
-            var prct = ((current / max) * 100),
-                color = undefined;
+                                range = config.maxAngle - config.minAngle;
+                                r = config.size / 2;
+                                pointerHeadLength = Math.round(r * config.pointerHeadLengthPercent);
 
-            if (prct < 34) {
-                color = "#3c9b33";
-            } else if (prct > 66) {
-                color = "#aa3f38"
-            } else {
-                color = "#c2bb48"
-            }
+                                // a linear scale that maps domain values to a percent from 0..1
+                                scale = d3.scale.linear()
+                                    .range([0,1])
+                                    .domain([config.minValue, config.maxValue]);
 
-            // Calculate draw angle
-            var drawAng = (((current * 180) / max) - 90);
+                                ticks = scale.ticks(config.majorTicks);
+                                tickData = d3.range(config.majorTicks).map(function() {return 1/config.majorTicks;});
 
-            var arc = d3.svg.arc()
-                .innerRadius(50)
-                .outerRadius(90)
-                .startAngle(-90 * (pi/180))
-                .endAngle(drawAng * (pi/180))
+                                arc = d3.svg.arc()
+                                    .innerRadius(r - config.ringWidth - config.ringInset)
+                                    .outerRadius(r - config.ringInset)
+                                    .startAngle(function(d, i) {
+                                        var ratio = d * i;
+                                        return deg2rad(config.minAngle + (ratio * range));
+                                    })
+                                    .endAngle(function(d, i) {
+                                        var ratio = d * (i+1);
+                                        return deg2rad(config.minAngle + (ratio * range));
+                                    });
+                            },
+                            centerTranslation = function() {
+                                return 'translate('+r +','+ r +')';
+                            },
+                            isRendered = function() {
+                                return (svg !== undefined);
+                            },
+                            render = function(newValue) {
+                                svg = d3.select($element[0])
+                                    .append('svg:svg')
+                                        .attr('class', 'gauge')
+                                        .attr('width', config.clipWidth)
+                                        .attr('height', config.clipHeight);
 
-            svg.append("path")
-                .attr("fill", color)
-                .attr("d", arc)
-                .attr("transform", "translate(130,100)")
+                                var centerTx = centerTranslation();
 
-            $scope.value = current;
+                                var arcs = svg.append('g')
+                                        .attr('class', 'arc')
+                                        .attr('transform', centerTx);
 
+                                arcs.selectAll('path')
+                                        .data(tickData)
+                                    .enter().append('path')
+                                        .attr('fill', function(d, i) {
+                                            return config.arcColorFn(d * i);
+                                        })
+                                        .attr('d', arc);
+
+                                var lg = svg.append('g')
+                                        .attr('class', 'label')
+                                        .attr('transform', centerTx);
+                                lg.selectAll('text')
+                                        .data(ticks)
+                                    .enter().append('text')
+                                        .attr('transform', function(d) {
+                                            var ratio = scale(d);
+                                            var newAngle = config.minAngle + (ratio * range);
+                                            return 'rotate(' +newAngle +') translate(0,' +(config.labelInset - r) +')';
+                                        })
+                                        .text(config.labelFormat);
+
+                                var lineData = [ [config.pointerWidth / 2, 0], 
+                                                [0, -pointerHeadLength],
+                                                [-(config.pointerWidth / 2), 0],
+                                                [0, config.pointerTailLength],
+                                                [config.pointerWidth / 2, 0] ];
+                                var pointerLine = d3.svg.line().interpolate('monotone');
+                                var pg = svg.append('g').data([lineData])
+                                        .attr('class', 'pointer')
+                                        .attr('transform', centerTx);
+
+                                pointer = pg.append('path')
+                                    .attr('d', pointerLine/*function(d) { return pointerLine(d) +'Z';}*/ )
+                                    .attr('transform', 'rotate(' +config.minAngle +')');
+
+                                update(newValue === undefined ? 0 : newValue);
+                            },
+                            update = function(newValue, newConfiguration) {
+                                if ( newConfiguration  !== undefined) {
+                                    configure(newConfiguration);
+                                }
+                                var ratio = scale(newValue);
+                                var newAngle = config.minAngle + (ratio * range);
+                                pointer.transition()
+                                    .duration(config.transitionMs)
+                                    .ease('elastic')
+                                    .attr('transform', 'rotate(' +newAngle +')');
+                            };
+
+                        that.configure = configure;
+
+                        that.isRendered = isRendered;
+                       
+                        that.render = render;
+                        
+                        that.update = update;
+
+                        configure(configuration);
+                        
+                        return that;
+                    };
+
+                powerGauge = gauge();
+                powerGauge.render();
+
+                // Need to add sort by time!!
+                $scope.$watch('data', function (aft, bef) {
+                    var sData = JSON.parse(aft);
+                    var x = sData.length - 1;
+                    powerGauge.update(sData[x].value.data.value);
+                }, true);
+
+            }, 0)           
         }
     }
-})
-
-
-
+});